Updated code with main latest, cleanned up code, and implemented Team tools test cases againts hcpt
