+++ test/terraform/teams_test.go
@@ -10,6 +10,14 @@ import (
 
 func TestListTeams(t *testing.T) {
 
+	// Guard: skip if the organization does not support team management.
+	client := tfeClient(t)
+	entitlements, err := client.Organizations.ReadEntitlements(t.Context(), tfeOrgName)
+	require.NoError(t, err, "Failed to read entitlements for organization %q", tfeOrgName)
+	if !entitlements.Teams {
+		t.Skipf("Organization %q does not have the Teams entitlement", tfeOrgName)
+	}
+
 	t.Run("list all teams", func(t *testing.T) {
 		s := newTestingSession(t)
 		defer s.Close()
@@ -26,7 +34,6 @@ func TestListTeams(t *testing.T) {
 		assert.NotEmpty(t, gjson.Get(resultText, "items.0.name").String(), "Tool call result should contain team names")
 		assert.NotEmpty(t, gjson.Get(resultText, "items.0.visibility").String(), "Tool call result should contain team visibility")
 		assert.True(t, gjson.Get(resultText, "items.0.users-count").Exists(), "Tool call result should contain user count field")
-
 	})
 
 	t.Run("filter by team name", func(t *testing.T) {
@@ -42,7 +49,6 @@ func TestListTeams(t *testing.T) {
 		require.NotEmpty(t, resultText, "Tool call result must not be empty")
 
 		assert.Equal(t, gjson.Get(resultText, "items.0.name").String(), "owners", "Filtered result should only contain the 'owners' team")
-
 	})
 
 	t.Run("filter by search query", func(t *testing.T) {
@@ -59,4 +65,52 @@ func TestListTeams(t *testing.T) {
 
 		assert.NotEqual(t, int(gjson.Get(resultText, "items.#").Int()), 0, "Search query should return at least one matching team")
 	})
+}
+
+func TestCreateTeam(t *testing.T) {
+
+	if tfeOrgName == "" {
+		t.Skip("You need to supply TFE_ORG_NAME to run this test")
+	}
+
+	s := newTestingSession(t)
+	defer s.Close()
+
+	client := tfeClient(t)
+	teamName := randomName("team-")
+	visibility := "organization"
+
+	// Guard: skip if the organization does not support team management.
+	entitlements, err := client.Organizations.ReadEntitlements(t.Context(), tfeOrgName)
+	require.NoError(t, err, "Failed to read entitlements for organization %q", tfeOrgName)
+
+	if !entitlements.Teams {
+		t.Skipf("Organization %q does not have the Teams entitlement", tfeOrgName)
+	}
+
+	result, resultText := callTool(t, s, "create_team", map[string]any{
+		"terraform_org_name": tfeOrgName,
+		"team_name":          teamName,
+		"visibility":         visibility,
+	})
+
+	require.False(t, result.IsError, "create_team returned an error: %s", resultText)
+	require.NotEmpty(t, resultText, "Tool call result must not be empty")
+
+	teamID := gjson.Get(resultText, "team_id").String()
+	require.NotEmpty(t, teamID, "Tool response should include a team ID")
+	defer client.Teams.Delete(t.Context(), teamID)
+
+	assert.Equal(t, teamName, gjson.Get(resultText, "team_name").String(),
+		"Tool response should report the created team name")
+
+	assert.Equal(t, visibility, gjson.Get(resultText, "visibility").String(),
+		"Tool response should report the requested visibility")
+
+	// Verify against the TFE API directly
+	createdTeam, err := client.Teams.Read(t.Context(), teamID)
+	require.NoError(t, err, "Team reported as created but produced an error when reading")
+	assert.Equal(t, teamName, createdTeam.Name, "Created team name does not match requested name")
+	assert.Equal(t, visibility, createdTeam.Visibility)
+	assert.Zero(t, createdTeam.UserCount)
 }
